Customer section done

+++ src/bot/customers/customer.service.ts
@@ -50,9 +50,9 @@ export class CustomerService {
 				const customer = await this.customersModel.findOne({
 					where: { user_id },
 				});
-				if (!customer?.status) {
+				if (customer && !customer?.status) {
 					const name = ctx.message!.text;
-					customer!.name = name;
+					customer.name = name;
 					await customer?.save();
 					await ctx.replyWithHTML("Enter your phone number:", {
 						...Markup.keyboard([
@@ -78,7 +78,7 @@ export class CustomerService {
 				const customerF = await this.masterCustomerModel.findOne({
 					where: { user_id: user_id },
 				});
-				if (customerF!.status == "feedback") {
+				if (customerF && customerF!.status == "feedback") {
 					customerF!.feedback = ctx.message.text;
 					customerF!.status = "complated";
 					await customerF!.save();
@@ -189,37 +189,56 @@ Metting status - ${meeting.status}
 			console.log(error);
 		}
 	}
-	generateTimeSlots(
-		startTime: string,
-		endTime: string,
-		stepMinutes: number,
+	formatTimeButtons(
+		input: string,
 		user_id: string
-	) {
-		const main: any[] = [];
-		const times: any[] = [];
-		const [startHour, startMinute] = startTime.split(":").map(Number);
-		const [endHour, endMinute] = endTime.split(":").map(Number);
-		let current = new Date();
-		current.setHours(startHour, startMinute, 0, 0);
-		const end = new Date();
-		end.setHours(endHour, endMinute, 0, 0);
-		while (current <= end) {
-			const hours = current.getHours().toString().padStart(2, "0");
-			const minutes = current.getMinutes().toString().padStart(2, "0");
-			times.push({
-				text: `${hours}:${minutes}`,
-				callback_data: `timeformetting_${hours}:${minutes}_${user_id}`,
-			});
-			if (times.length === 4) {
-				main.push([...times]);
-				times.length = 0;
+	): { text: string; callback_data: string }[][] {
+		const parts = input.split("||").filter(Boolean);
+		const result: { text: string; callback_data: string }[][] = [];
+		const row: { text: string; callback_data: string }[] = [];
+		for (const part of parts) {
+			const [timeRange, status] = part.split("|");
+			// let emoji = status === "t" ? "✅" : "❌";
+			let emoji = "";
+			let tmp = "";
+			if (status == "t") {
+				emoji = "✅";
+				tmp = "1";
+			} else {
+				emoji = "❌";
+				tmp = "0";
 			}
-			current.setMinutes(current.getMinutes() + stepMinutes);
+			const button = {
+				text: `${timeRange} ${emoji}`,
+				callback_data: `timeformetting_${timeRange}_${user_id}_${tmp}`,
+			};
+			row.push(button);
+			if (row.length === 2) {
+				result.push([...row]);
+				row.length = 0;
+			}
 		}
-		if (times.length > 0) {
-			main.push([...times]);
+		if (row.length) {
+			result.push([...row]);
 		}
-		return main;
+		return result;
+	}
+
+	updateTimeSlotStatus(
+		input: string,
+		fromTime: string,
+		toTime: string,
+		newStatus: "t" | "f"
+	): string {
+		const parts = input.split("||").filter(Boolean);
+		const updated = parts.map((part) => {
+			const [timeRange, status] = part.split("|");
+			if (timeRange === `${fromTime}-${toTime}`) {
+				return `${timeRange}|${newStatus}`;
+			}
+			return part;
+		});
+		return updated.join("||");
 	}
 
 	async onMettingStepWithWho(ctx: Context) {
@@ -239,20 +258,19 @@ Metting status - ${meeting.status}
 				const master = await this.mastersModel.findOne({
 					where: { user_id: master_id },
 				});
-				const startTime = master?.work_start_time!;
-				const endTime = master?.work_end_time!;
-				const step = Number(master?.avgtime_for_custommer.split(" ")[0]);
-				const buttons = this.generateTimeSlots(
-					startTime,
-					endTime,
-					step + 20,
-					user_id
-				);
+				const buttons = this.formatTimeButtons(master!.times, user_id);
 				await ctx.replyWithHTML(`Choose the metting time:`, {
 					reply_markup: {
 						inline_keyboard: buttons,
 					},
 				});
+			} else {
+				await this.masterCustomerModel.create({
+					user_id,
+					master_id,
+					last_state: "withwho",
+				});
+				this.onMettingStepWithWho(ctx)
 			}
 		} catch (error) {
 			console.log(error);
@@ -276,6 +294,29 @@ Metting status - ${meeting.status}
 
 		return result;
 	}
+	isTimeSlotBusy(busyRanges: string, checkRange: string): boolean {
+		if (!busyRanges || busyRanges == "") {
+			return false;
+		}
+		const checkStart = checkRange.split("-")[0];
+		const checkEnd = checkRange.split("-")[1];
+
+		const busyList = busyRanges.split("||").filter(Boolean);
+
+		for (const range of busyList) {
+			const [start, end] = range.split("-");
+
+			if (
+				(checkStart >= start && checkStart < end) ||
+				(checkEnd > start && checkEnd <= end) ||
+				(checkStart <= start && checkEnd >= end)
+			) {
+				return true;
+			}
+		}
+
+		return false;
+	}
 
 	async onMettingStepTime(ctx: Context) {
 		try {
@@ -283,11 +324,46 @@ Metting status - ${meeting.status}
 			const contextMessage = ctx.callbackQuery!["message"];
 			const user_id = contextAction.split("_")[2];
 			const time = contextAction.split("_")[1];
+			const st = contextAction.split("_")[3];
 			const makker = await this.masterCustomerModel.findOne({
 				where: { user_id, last_state: "time" },
 			});
-			if (makker) {
+			const master = await this.mastersModel.findOne({
+				where: { user_id: makker?.master_id },
+			});
+			const customer = await this.customersModel.findOne({
+				where: { user_id },
+			});
+			if (st == "0") {
+				const buttons = this.formatTimeButtons(master!.times, user_id);
+				await ctx.replyWithHTML(
+					`Master is busy on this time, please choose another metting time:`,
+					{
+						reply_markup: {
+							inline_keyboard: buttons,
+						},
+					}
+				);
+			} else if (
+				makker &&
+				!this.isTimeSlotBusy(
+					customer!.times,
+					`${time.split(":")[0] - time.split(":")[1]}`
+				)
+			) {
 				makker.time = time;
+				let cstTime = customer!.times;
+				cstTime += `||${time.split(":")[0] - time.split(":")[1]}`;
+				customer!.times = cstTime;
+				await customer?.save();
+				const inputs = master!.times;
+				master!.times = this.updateTimeSlotStatus(
+					inputs,
+					time.split("-")[0],
+					time.split("-")[1],
+					"f"
+				);
+				await master?.save();
 				makker.last_state = "date";
 				await makker.save();
 				await ctx.deleteMessage(contextMessage?.message_id);
@@ -302,6 +378,16 @@ Metting status - ${meeting.status}
 						inline_keyboard: buttons,
 					},
 				});
+			} else {
+				const buttons = this.formatTimeButtons(master!.times, user_id);
+				await ctx.replyWithHTML(
+					`You have a metting on this time, please choose another metting time:`,
+					{
+						reply_markup: {
+							inline_keyboard: buttons,
+						},
+					}
+				);
 			}
 		} catch (error) {
 			console.log(error);
@@ -523,6 +609,190 @@ Metting status - ${makker.status}
 					]).resize(),
 				});
 			}
+		} catch (error) {
+			console.log(error);
+		}
+	}
+	async onServices(ctx: Context) {
+		try {
+			await ctx.replyWithHTML("Select the desired section:", {
+				...Markup.keyboard([
+					["<Hairdresser>", "<Beauty Salon>"],
+					["<Jewelry Workshop>", "<Watchmaker>"],
+					["<Shoe Workshop>"],
+					["<Back to Main>"],
+				]).resize(),
+			});
+		} catch (error) {
+			console.log(error);
+		}
+	}
+	async onSections(ctx: Context, section: string) {
+		try {
+			if (section == "Back to Main") {
+				await ctx.replyWithHTML(`Choose menu:`, {
+					...Markup.keyboard([
+						["My meetings", "Make an appointment"],
+						["Services"],
+					]).resize(),
+				});
+				return;
+			}
+			const masters = await this.mastersModel.findAll({
+				where: { section },
+			});
+			if (!masters.length) {
+				await ctx.replyWithHTML(`No masters found in section ${section}:`, {
+					...Markup.keyboard([
+						["<Hairdresser>", "<Beauty Salon>"],
+						["<Jewelry Workshop>", "<Watchmaker>"],
+						["<Shoe Workshop>"],
+						["<Back to Main>"],
+					]).resize(),
+				});
+			}
+			let str = "";
+			for (const master of masters) {
+				str += `1. Name - ${master.name}
+2. Phone number - ${master.phone_number}
+3. Workshop name - ${master.workshop_name ? master.workshop_name : "Not included"}
+4. Workshop address - ${master.address ? master.address : "Not included"}
+5. Workshop target - ${master.target_for_address ? master.target_for_address : "Not included"}
+6. Location - ${master.location}
+7. Time to start work - ${master.work_start_time}
+8. Time to finish work - ${master.work_end_time}
+9. Avarage time spent per customer - ${master.avgtime_for_custommer}
+10. Avarage masters mark - ${master.rating == 5 ? "⭐⭐⭐⭐⭐" : master.rating >= 4 && master.rating < 5 ? "⭐⭐⭐⭐" : master.rating > 3 && master.rating < 4 ? "⭐⭐⭐" : "⭐⭐"}, (${Number(master.rating).toFixed(2)})
+							`;
+			}
+			await ctx.replyWithHTML(str, {
+				...Markup.keyboard([
+					["Get closest by location"],
+					["Get by rating"],
+					["<Back>"],
+				]).resize(),
+			});
+		} catch (error) {
+			console.log(error);
+		}
+	}
+
+	deg2rad(deg: number): number {
+		return deg * (Math.PI / 180);
+	}
+	getDistanceFromLatLonInKm(
+		lat1: number,
+		lon1: number,
+		lat2: number,
+		lon2: number
+	): number {
+		const R = 6371;
+		const dLat = this.deg2rad(lat2 - lat1);
+		const dLon = this.deg2rad(lon2 - lon1);
+		const a =
+			Math.sin(dLat / 2) ** 2 +
+			Math.cos(this.deg2rad(lat1)) *
+				Math.cos(this.deg2rad(lat2)) *
+				Math.sin(dLon / 2) ** 2;
+		const c = 2 * Math.atan2(Math.sqrt(a), Math.sqrt(1 - a));
+		return R * c;
+	}
+	async querySendCustomerLoc(ctx: Context) {
+		try {
+			const user_id = String(ctx.from!.id);
+			await this.customersModel.update(
+				{ actions: "sendlocation" },
+				{ where: { user_id: String(user_id) } }
+			);
+			await ctx.replyWithHTML(`Send your location: `, {
+				...Markup.keyboard([
+					Markup.button.locationRequest("Send location"),
+					"< Back",
+				]).resize(),
+			});
+		} catch (error) {
+			console.log(error);
+		}
+	}
+	async findClosestByLocation(
+		ctx: Context,
+		lat: any,
+		lon: any,
+		userId: string
+	) {
+		try {
+			const masters = await this.mastersModel.findAll();
+			const distances: any = [];
+			await this.masterCustomerModel.create({
+				user_id: userId,
+				last_state: "withwho",
+			});
+
+			let counter = 0;
+			for (const master of masters) {
+				const dist = this.getDistanceFromLatLonInKm(
+					lat,
+					lon,
+					Number(master.location.split("|")[0]),
+					Number(master.location.split("|")[1])
+				);
+				counter++;
+				const distance: any = [];
+				distance.push({
+					text: `${counter}. ${master.name} ${master.section} ${(dist * 1000).toFixed(2)} m.`,
+					callback_data: `makeapp_${master.user_id}_${userId}`,
+				});
+				distances.push(distance);
+			}
+			await ctx.replyWithHTML("Closest masters:", {
+				reply_markup: {
+					inline_keyboard: distances,
+				},
+			});
+		} catch (error) {
+			console.log(error);
+		}
+	}
+	async backGetCloseLoc(ctx: Context) {
+		await this.customersModel.update(
+			{ actions: "" },
+			{ where: { user_id: String(ctx.from!.id) } }
+		);
+		await ctx.replyWithHTML(`Choose:`, {
+			...Markup.keyboard([
+				["Get closest by location"],
+				["Get by rating"],
+				["<Back>"],
+			]).resize(),
+		});
+	}
+	async getByRating(ctx: Context) {
+		try {
+			const userId = String(ctx.from!.id);
+			const masters = await this.mastersModel.findAll();
+			for (let i = 0; i < masters.length - 1; i++) {
+				for (let j = 0; j < masters.length - i - 1; j++) {
+					if (masters[j].rating < masters[j + 1].rating) {
+						[masters[j], masters[j + 1]] = [masters[j + 1], masters[j]];
+					}
+				}
+			}
+			const btns: any = [];
+			let counter = 0;
+			for (const master of masters) {
+				counter++;
+				const tmp: any = [];
+				tmp.push({
+					text: `${counter}. ${master.name} ${master.section} ${master.rating == 5 ? "⭐⭐⭐⭐⭐" : master.rating >= 4 && master.rating < 5 ? "⭐⭐⭐⭐" : master.rating > 3 && master.rating < 4 ? "⭐⭐⭐" : "⭐⭐"}, (${Number(master.rating).toFixed(2)})`,
+					callback_data: `makeapp_${master.user_id}_${userId}`,
+				});
+				btns.push(tmp);
+			}
+			await ctx.replyWithHTML("Choose:", {
+				reply_markup: {
+					inline_keyboard: btns,
+				},
+			});
 		} catch (error) {
 			console.log(error);
 		}
